requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
Django==4.1.5
django-bootstrap-v5==1.0.11
django-filter==22.1
django-tables2==2.5.2
numpy==1.24.2
pandas==1.5.3
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
